api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.1",
    "azure-cosmos>=4.14.0",
    "azure-identity>=1.25.1",
    "azure-storage-blob>=12.27.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.119.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
    "types-aiofiles>=25.1.0.20251011",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "pytest-cov>=4.0.0",
]
